Added Orientation option
Changed bits of code all around to handle that properly. Especially in the
function ComputeDeltas() and HandleScrolling(). Those two functions convert
(using helper functions that I've added) the hw->x and hw->y coordinates to
oriented_x and oriented_y coordinates. The angle used for circular scrolling
do not use oriented coordinates as i've taught it was not necessary, but I
haven't tested it though (i don't use circular scrolling).

Moreover, in the function HandleState(), the variable edge contains how the
oriented edge. So the functions that use the edge are automatically oriented.
This is useful for HandleTap(). The buttons in the corners of the pads are
oriented without effort.