Added DontReportSize option
This option prevent synaptics from reporting the size of the trackpad to Xorg,
this is useful to prevent Xorg from changing the speed of one of the axis
depending of the size of the pad. This is especially useful since when we use
the orientation option, we virtually change the trackpad dimentions (orienting
it), so the modification Xorg does on the axis afterwards are wrong, and makes
the orientation unusable.

Perhaps there is a way to change the trackpad dimentions when we change
orientation, but I don't know enough of the Worg internals to do that.
Moreover, this option is still useful since the speed changes Xorg makes on my
y axis makes me usable to reach some pixels on the screen (the slightest move I
make on the y axis with my pad results in a movement of two pixels). In one
want accurate positioning with the pad, this option could be useful.