Added HorizSpeed and VertSpeed options
Using the option DontReportSize, the two axis always have the same speed. This
may be undesirale in some cases when physically, the trackpad has an axis more
precise than the other (in my case, the x axis is two times more precise than
the y axis, so I have to speed up the y axis speed).