Unimportant changes
I changed somewhere a priv->synpara->... in para->...

I also put the code that emulate middle and right clicks using the left click
button (a code i've also written) before after the emulation of two-finger (on
pads that don't report multiple fingers). So this code can be useful on such
pads (even if it may not be really usable, i don't know how this detection of
multiple fingers is accurate).